+++ examples/browser-http/server/main.go
@@ -1,3 +1,5 @@
+//+build js
+
 package main
 
 import (
